% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{as.named.vector}
\alias{as.named.vector}
\title{as.named.vector}
\usage{
as.named.vector(YourDataFrameWithRownames, WhichDimNames = 1)
}
\arguments{
\item{WhichDimNames}{Set to 1 for row names of a selected colun, set to 2 for column names if you selected a row.}

\item{A}{single column (or row) of a dataframe}
}
\description{
as.named.vector
}
\examples{
as.named.vector(YourDataFrameWithRownames[, 1], WhichDimNames = 1)
}

