% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{descriptor_roxy.headerOnly}
\alias{descriptor_roxy.headerOnly}
\title{descriptor_roxy.headerOnly}
\usage{
descriptor_roxy.headerOnly(function_to_parse)
}
\arguments{
\item{function_to_parse}{name of the function in the memory space to parse to clipboard}
}
\description{
Description ...
}
\examples{
descriptor_roxy(ls_funs = funnames2, OutputFile = Package_FnP)
}

