% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{descriptor_roxy}
\alias{descriptor_roxy}
\title{descriptor_roxy}
\usage{
descriptor_roxy(ls_funs, OutputFile)
}
\arguments{
\item{ls_funs}{list of functions to include in the package: list names should be exact function names in the NAMESPACE,list elements should be the functions.}

\item{OutputFile}{Full path and file name of the file the Roxygen-ready script is written to}
}
\description{
Description ...
}
\examples{
descriptor_roxy(ls_funs = funnames2, OutputFile = Package_FnP)
}

