% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{descriptor_roxy.singleFunction}
\alias{descriptor_roxy.singleFunction}
\title{descriptor_roxy.singleFunction}
\usage{
descriptor_roxy.singleFunction(function_to_parse, OutputFile)
}
\arguments{
\item{function_to_parse}{name of the function in the memory space to parse into a new file}

\item{OutputFile}{Full path and file name of the file the Roxygen-ready script is written to}
}
\description{
Description ...
}
\examples{
descriptor_roxy(ls_funs = funnames2, OutputFile = Package_FnP)
}

