% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{descriptor_roxy_old}
\alias{descriptor_roxy_old}
\title{descriptor_roxy_old}
\usage{
descriptor_roxy_old(FuncNames, outFile)
}
\arguments{
\item{FuncNames}{a vector of function names (of functions in the memory space) to be annotated}

\item{outFile}{Filename and path where the annotated version will be written out}
}
\description{
Compile and write out all functions documentation in a script for Roxygen
}
\examples{
descriptor_roxy_old (FuncNames =  , outFile)
}

