% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{funnotator_RoxygenReady.headerOnly}
\alias{funnotator_RoxygenReady.headerOnly}
\title{funnotator_RoxygenReady.headerOnly}
\usage{
funnotator_RoxygenReady.headerOnly(function_to_parse)
}
\arguments{
\item{function_to_parse}{Name of the function with the arguments of interest.}
}
\description{
Compile a single functions documentation for Roxygen
}
\examples{
funnotator_RoxygenReady.headerOnly (function_to_parse =  )
}

