% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{funnotator_RoxygenReady}
\alias{funnotator_RoxygenReady}
\title{funnotator_RoxygenReady}
\usage{
funnotator_RoxygenReady(FileWithFunctions,
  outFile = kollapse(FileWithFunctions, ".annot.R", print = F))
}
\arguments{
\item{FileWithFunctions}{An .R file containing all the function definitions of your library}

\item{outFile}{Filename and path where the annotated version will be written out}
}
\description{
Compile and write out all functions documentation in a script for Roxygen (Extended version)
}
\examples{
funnotator_RoxygenReady (FileWithFunctions =  , outFile = kollapse(FileWithFunctions, ".annot.R", print = F))
}

