% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{funnotator_RoxygenReady.singleFunction}
\alias{funnotator_RoxygenReady.singleFunction}
\title{funnotator_RoxygenReady.singleFunction}
\usage{
funnotator_RoxygenReady.singleFunction(function_to_parse, outFile)
}
\arguments{
\item{function_to_parse}{Name of the function with the arguments of interest.}

\item{outFile}{Filename and path where the annotated version will be written out}
}
\description{
Compile and write out a single functions documentation for Roxygen
}
\examples{
funnotator_RoxygenReady.singleFunction (function_to_parse =  , outFile = )
}

