% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{kollapse}
\alias{kollapse}
\title{kollapse}
\usage{
kollapse(..., print = T)
}
\description{
paste0 values and string to one string. It also prints the results (good for a quick check)
}
\examples{
kollapse("Hello ", "you ", 3, ", " , 11, " year old kids.")
}

