% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{RoxygenReady}
\alias{RoxygenReady}
\title{RoxygenReady.FileToFile}
\usage{
RoxygenReady(FileWithFunctions, outFile = kollapse(FileWithFunctions,
  ".annot.R", print = F, overview = T))
}
\arguments{
\item{FileWithFunctions}{An .R file containing all the function definitions of your library}

\item{outFile}{Filename and path where the annotated version will be written out}

\item{overview}{If true, a function overview .md file is created next to your input file. See rr_function_overview_document() for details.}
}
\description{
Read in a file, annotate and write out all functions documentation with Roxygen skeleton (FileToFile)
}

