% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{rr_extract__all_descriptions_from_comment}
\alias{rr_extract__all_descriptions_from_comment}
\title{rr_extract__all_descriptions_from_comment}
\usage{
rr_extract__all_descriptions_from_comment(inFile)
}
\arguments{
\item{inFile}{input file with function definitions to be scanned}
}
\description{
Scan a script for (descriptive) comments in the first line of each function's definition.
}
\examples{
rr_extract__all_descriptions_from_comment (inFile =  )
}

