% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{rr_extract_all_descriptions_from_comment_annot}
\alias{rr_extract_all_descriptions_from_comment_annot}
\title{rr_extract_all_descriptions_from_comment_annot}
\usage{
rr_extract_all_descriptions_from_comment_annot(inFile)
}
\arguments{
\item{inFile}{input file with function definitions to be scanned}
}
\description{
Scan a script for (descriptive) comments in the 3d line of each function's Roxygen skeleton.
}
\examples{
rr_extract_all_descriptions_from_comment_annot (inFile =  )
}

