% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{toClipboard}
\alias{toClipboard}
\title{toClipboard}
\usage{
toClipboard(x, sep = "\\t", header = FALSE, row.names = FALSE,
  col.names = F)
}
\arguments{
\item{x}{string}

\item{sep}{the field separator string. Values within each row of x are separated by this string.}

\item{header}{TRUE if the variable has a header line. FALSE by default.}

\item{row.names}{either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.}

\item{col.names}{either a logical value indicating whether the column names of x are to be written along with x, or a character vector of column names to be written. See the section on <U+2018>CSV files<U+2019> for the meaning of col.names = NA.}
}
\description{
Copy an R-object to your clipboard on OS X.
}
\examples{
toClipboard(11)
}

