% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RoxygenReady.R
\name{toClipboard}
\alias{toClipboard}
\title{toClipboard}
\usage{
toClipboard(x, sep = "\\t", header = FALSE, row.names = FALSE,
  col.names = F)
}
\description{
# Copy an R-object to your clipboard on OS X.
}
\examples{
toClipboard(11)
}

